+++ src/Landing/Section02.jsx
@@ -1,18 +1,18 @@
-// 랜딩 페이지 섹션 3 컴포넌트
+// 랜딩 페이지 섹션 2 컴포넌트
 import { motion } from 'framer-motion';
 import { useInView } from 'react-intersection-observer';
-import bgImage03 from '../../assets/images/landing_bg03.png';
-import PhonePage02 from '../../assets/images/PhonePage02.png';
+import bgImage02 from '../assets/images/landing_bg02.png';
+import PhonePage01 from '../assets/images/PhonePage01.png';
 
-function Section3() {
+function Section2() {
   const { ref, inView } = useInView({ threshold: 0.2, triggerOnce: false });
 
   return (
     <motion.section
       ref={ref}
-      className="landing-section section3"
+      className="landing-section section2"
       style={{
-        backgroundImage: `url(${bgImage03})`,
+        backgroundImage: `url(${bgImage02})`,
         backgroundSize: 'cover',
         backgroundPosition: 'center',
       }}
@@ -26,7 +26,7 @@ function Section3() {
           animate={inView ? { opacity: 1, y: 0 } : { opacity: 0, y: 30 }}
           transition={{ duration: 0.5, delay: 0.1 }}
         >
-          <span className="yellow-highlight">이달의 아티스트</span>
+          <span className="yellow-highlight">후원하기</span>
         </motion.h3>
 
         <motion.h2
@@ -34,14 +34,14 @@ function Section3() {
           animate={inView ? { opacity: 1, y: 0 } : { opacity: 0, y: 30 }}
           transition={{ duration: 0.5, delay: 0.3 }}
         >
-          내 아티스트에게 1등의
+          좋아하는 아이돌에게
           <br />
-          영예를 선물하세요
+          쉽게 조공해보세요
         </motion.h2>
 
         <motion.img
-          src={PhonePage02}
-          alt="이달의 아티스트 기능을 보여주는 앱 화면"
+          src={PhonePage01}
+          alt="아이돌 후원 기능을 보여주는 앱 화면"
           className="PhonePage"
           initial={{ opacity: 0, y: 30 }}
           animate={inView ? { opacity: 1, y: 0 } : { opacity: 0, y: 30 }}
@@ -52,4 +52,4 @@ function Section3() {
   );
 }
 
-export default Section3;
+export default Section2;
