+++ src/auth/validateAllFields.jsx
