--- src/components/CreditChargeModalBody.module.css
@@ -1,35 +1,0 @@
-.radioWrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  margin-bottom: 20px;
-}
-.radioBox {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  box-sizing: border-box;
-  width: 295px;
-  border: 1px solid white;
-  border-radius: 8px;
-  padding: 16px 20px;
-  transition: 0.1s;
-  cursor: pointer;
-}
-.radioLeft {
-  display: flex;
-  align-items: center;
-}
-.selected {
-  border-color: var(--mantine-color-brand-0);
-}
-.creditText {
-  color: var(--mantine-color-gray-1);
-  font-weight: 700;
-  font-size: 20px;
-  line-height: 26px;
-  margin-top: 2px;
-}
-.creditSelected {
-  color: white;
-}
