--- src/components/Donation.jsx
@@ -1,6 +1,0 @@
-import IdolSponsor from './IdolSponsor';
-
-function Donation() {
-  return <div>{/* <IdolSponsor /> */}</div>;
-}
-export default Donation;
