--- src/components/config.js
@@ -1,3 +1,0 @@
-const CLIENT_STATE_KEYS = {
-  credits: 'credits',
-};
