--- src/settings/Cursor.jsx
@@ -1,5 +1,0 @@
-function Cursor() {
-  return;
-}
-
-export default Cursor;
